# api/openapi.yaml
openapi: 3.1.0
info:
  title: Collections Agent API
  version: 1.0.0
  description: REST API for the AI Receivables Collections Agent console, the bank webhook and admin controls. Serves the
    stories named in each operation's x-story-ids. Money fields are integer paise named *_paise (INR only); every amount shown
    to a user is formatted from these at the edge.
  contact:
    name: HACK-001 team
...
servers:
- url: http://localhost:8000/api/v1
  description: local Docker Compose
security:
- demoRole: []
...
paths:
...
  /runs:
    post:
      operationId: startDailyRun
      summary: 'Run now: enqueue today''s daily run'
      tags:
      - runs
      x-story-ids:
      - US-01-001
      x-roles: admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunStart'
      responses:
        '202':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunBatch'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      operationId: listRuns
      summary: Recent runs
      tags:
      - runs
      x-story-ids:
      - US-01-006
      x-roles: admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                - page
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AgentRun'
                  page:
                    $ref: '#/components/schemas/Page'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /runs/{id}:
    get:
      operationId: getRun
      summary: One run with its steps
      tags:
      - runs
      x-story-ids:
      - US-00-008
      x-roles: viewer, collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRun'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /messages:
    get:
      operationId: listMessages
      summary: Approval queue and message lists
      tags:
      - messages
      x-story-ids:
      - US-00-009
      x-roles: viewer, collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      - name: filter[status]
        in: query
        required: false
        schema:
          type: string
          enum:
          - draft
          - pending_approval
          - approved
          - rejected
          - sent
          - failed
      - name: filter[customer_id]
        in: query
        required: false
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                - page
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
                  page:
                    $ref: '#/components/schemas/Page'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /messages/{id}:
    get:
      operationId: getMessage
      summary: One message with its guardrail report
      tags:
      - messages
      x-story-ids:
      - US-00-009
      x-roles: viewer, collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      operationId: editMessage
      summary: Edit text; re-runs guardrails; If-Match carries the version
      tags:
      - messages
      x-story-ids:
      - US-00-010
      x-roles: collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageEdit'
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /messages/{id}/approve:
    post:
      operationId: approveMessage
      summary: Approve (idempotent; refuses unverified or stale versions)
      tags:
      - messages
      x-story-ids:
      - US-00-009
      - US-00-010
      x-roles: collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /messages/{id}/reject:
    post:
      operationId: rejectMessage
      summary: Reject with a reason
      tags:
      - messages
      x-story-ids:
      - US-00-009
      x-roles: collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reject'
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /messages/{id}/resend:
    post:
      operationId: resendMessage
      summary: Manual resend of a failed or unconfirmed message
      tags:
      - messages
      x-story-ids:
      - US-00-011
      x-roles: collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - $ref: '#/components/parameters/IdempotencyKey'
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /messages/approve-batch:
    post:
      operationId: approveMessagesBatch
      summary: 'Assisted mode: approve passed drafts in one action'
      tags:
      - messages
      x-story-ids:
      - US-01-016
      x-roles: collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchApprove'
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                - page
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
                  page:
                    $ref: '#/components/schemas/Page'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /replies:
    post:
      operationId: createReply
      summary: Simulate a customer reply to a sent message; classification runs
      tags:
      - replies
      x-story-ids:
      - US-03-001
      - US-00-012
      x-roles: collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplyCreate'
      responses:
        '201':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reply'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /replies/{id}:
    get:
      operationId: getReply
      summary: One reply with classification
      tags:
      - replies
      x-story-ids:
      - US-00-012
      x-roles: viewer, collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reply'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    demoRole:
      type: apiKey
      in: header
      name: X-Demo-Role
      description: admin, collector or viewer
    bankSignature:
      type: apiKey
      in: header
      name: X-Bank-Signature
      description: hex HMAC-SHA256 of '<X-Bank-Timestamp>.<raw body>' with BANK_WEBHOOK_SECRET
  parameters:
    ApiVersion:
      name: X-API-Version
      in: header
      required: false
      schema:
        type: string
        example: '2026-09-30'
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: The message version the client saw; a different version returns 409 STALE_DRAFT
      schema:
        type: string
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
  headers:
    XRequestId:
      schema:
        type: string
      description: Matches error.request_id and the server log line
  responses:
    BadRequest:
      description: Malformed request
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: BAD_REQUEST
              message: Malformed request
              request_id: req_01J9Z6
    Unauthorized:
      description: Not signed in
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHORIZED
              message: Not signed in
              request_id: req_01J9Z6
    Forbidden:
      description: Signed in but not allowed
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: FORBIDDEN
              message: Signed in but not allowed
              request_id: req_01J9Z6
    NotFound:
      description: Not found
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: NOT_FOUND
              message: Not found
              request_id: req_01J9Z6
    Conflict:
      description: State or idempotency conflict
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: STALE_DRAFT
              message: State or idempotency conflict
              request_id: req_01J9Z6
    UnprocessableEntity:
      description: Valid shape, invalid content
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: AMOUNT_MISMATCH
              message: Valid shape, invalid content
              request_id: req_01J9Z6
...
    InternalError:
      description: Unexpected error
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: INTERNAL
              message: Unexpected error
              request_id: req_01J9Z6
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Stable UPPER_SNAKE code, e.g. AMOUNT_MISMATCH, NOT_APPROVED, STALE_DRAFT
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  reason:
                    type: string
                required:
                - field
                - reason
            request_id:
              type: string
          required:
          - code
          - message
          - request_id
      required:
      - error
      example:
        error:
          code: AMOUNT_MISMATCH
          message: ₹4,50,000 is not the ledger amount for INV-1021
          details:
          - field: body
            reason: INV-1021 is 40000000 paise
          request_id: req_01J9Z6
    Page:
      type: object
      properties:
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
        has_more:
          type: boolean
      required:
      - next_cursor
      - has_more
      example:
        next_cursor: eyJpZCI6IjAxOTIifQ
        has_more: false
...
    GuardrailCheck:
      type: object
      properties:
        check:
          type: string
        token:
          type: string
        expected:
          anyOf:
          - type: string
          - type: 'null'
        ok:
          type: boolean
        code:
          anyOf:
          - type: string
          - type: 'null'
      required:
      - check
      - token
      - expected
      - ok
      - code
      example:
        check: amount
        token: ₹4,00,000
        expected: INV-1021 40000000
        ok: true
        code: null
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        agent_run_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
        kind:
          type: string
          enum:
          - reminder
          - followup
          - statement
          - dispute_ack
        channel:
          type: string
          enum:
          - email
          - whatsapp
          - voice
        tone:
          type: string
          enum:
          - gentle
          - firm
          - final
        status:
          type: string
          enum:
          - draft
          - pending_approval
          - approved
          - rejected
          - sent
          - failed
        subject:
          type: string
        body:
          type: string
        version:
          type: integer
          minimum: 1
        verified:
          type: boolean
          description: True when the current version passed guardrails
        guardrail_report:
          type: array
          items:
            $ref: '#/components/schemas/GuardrailCheck'
        invoice_ids:
          type: array
          items:
            type: string
            format: uuid
        rejected_reason:
          anyOf:
          - type: string
          - type: 'null'
        sent_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
        last_error:
          anyOf:
          - type: string
          - type: 'null'
        created_at:
          type: string
          format: date-time
      required:
      - id
      - customer_id
      - agent_run_id
      - kind
      - channel
      - tone
      - status
      - subject
      - body
      - version
      - verified
      - guardrail_report
      - invoice_ids
      - rejected_reason
      - sent_at
      - last_error
      - created_at
      example:
        id: 0192b1c2-7f3a-7c4e-9a1b-2c3d4e5f6a7b
        customer_id: 0192b1c2-7f3a-7c4e-9a1b-2c3d4e5f6a7b
        agent_run_id: 0192b1c2-7f3a-7c4e-9a1b-2c3d4e5f6a7b
        kind: reminder
        channel: email
        tone: firm
        status: pending_approval
        subject: Overdue invoices for ABC Distributors
        body: Dear ABC Distributors, ... INV-1021 | ₹4,00,000 | due 11 Sep 2026 ... Total ₹7,50,000
        version: 1
        verified: true
        guardrail_report: []
        invoice_ids:
        - 0192b1c2-7f3a-7c4e-9a1b-2c3d4e5f6a7b
        rejected_reason: null
        sent_at: null
        last_error: null
        created_at: '2026-09-30T03:31:00Z'
    MessageEdit:
      type: object
      properties:
        body:
          type: string
          minLength: 1
        subject:
          type: string
          minLength: 1
        channel:
          anyOf:
          - type: string
            enum:
            - email
            - whatsapp
          - type: 'null'
      required:
      - body
      - subject
      example:
        subject: Overdue invoices for ABC Distributors
        body: Dear ABC Distributors, ...
    Reject:
      type: object
      properties:
        reason:
          type: string
          minLength: 1
      required:
      - reason
      example:
        reason: Tone too firm for a first reminder
    BatchApprove:
      type: object
      description: Each draft at the version the collector saw; a newer version refuses the whole batch (STALE_DRAFT).
      properties:
        messages:
          type: array
          minItems: 1
          maxItems: 50
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              version:
                type: integer
                minimum: 1
            required:
            - id
            - version
      required:
      - messages
      example:
        messages:
        - id: 0192b1c2-7f3a-7c4e-9a1b-2c3d4e5f6a7b
          version: 1
    ReplyCreate:
      type: object
      properties:
        message_id:
          type: string
          format: uuid
        body:
          type: string
          minLength: 1
          maxLength: 5000
      required:
      - message_id
      - body
      example:
        message_id: 0192b1c2-7f3a-7c4e-9a1b-2c3d4e5f6a7b
        body: We can pay ₹3 lakh on October 5 and the remaining amount later.
...
    AgentRun:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        run_date:
          type: string
          format: date
        trigger:
          type: string
          enum:
          - scheduled
          - manual
          - reply
          - payment
        status:
          type: string
          enum:
          - running
          - finished
        outcome:
          anyOf:
          - type: string
            enum:
            - WAIT_FOR_APPROVAL
            - ESCALATED
            - STOPPED_LIMIT
            - NO_ACTION
            - FAILED
          - type: 'null'
        tool_call_count:
          type: integer
          minimum: 0
          maximum: 4
        action:
          anyOf:
          - type: string
          - type: 'null'
        channel:
          anyOf:
          - type: string
          - type: 'null'
        tone:
          anyOf:
          - type: string
          - type: 'null'
        reason:
          anyOf:
          - type: string
          - type: 'null'
        started_at:
          type: string
          format: date-time
        finished_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
        steps:
          type: array
          items:
            $ref: '#/components/schemas/AgentStep'
      required:
      - id
      - customer_id
      - run_date
      - trigger
      - status
      - outcome
      - tool_call_count
      - action
      - channel
      - tone
      - reason
      - started_at
      - finished_at
      - steps
      example:
        id: 0192b1c2-7f3a-7c4e-9a1b-2c3d4e5f6a7b
        customer_id: 0192b1c2-7f3a-7c4e-9a1b-2c3d4e5f6a7b
        run_date: '2026-09-30'
        trigger: manual
        status: finished
        outcome: WAIT_FOR_APPROVAL
        tool_call_count: 2
        action: send_reminder
        channel: email
        tone: firm
        reason: 3 overdue invoices, 1 missed promise
        started_at: '2026-09-30T03:30:00Z'
        finished_at: '2026-09-30T03:30:07Z'
        steps: []
    RunStart:
      type: object
      properties:
        customer_ids:
          type: array
          items:
            type: string
            format: uuid
      required: []
      example:
        customer_ids: []
    RunBatch:
      type: object
      properties:
        run_date:
          type: string
          format: date
        job_id:
          anyOf:
          - type: integer
          - type: 'null'
        queued:
          type: boolean
        run_ids:
          type: array
          items:
            type: string
            format: uuid
      required:
      - run_date
      - job_id
      - queued
      - run_ids
      example:
        run_date: '2026-09-30'
        job_id: null
        queued: false
        run_ids:
        - 0192b1c2-7f3a-7c4e-9a1b-2c3d4e5f6a7b
      description: 'No customer_ids: today''s daily run is queued (job_id). customer_ids: those customers run now, in the
        request (run_ids).'
...